created models folder and removed migrations from git
