+++ .gitignore
@@ -3,3 +3,6 @@
 dh-param/dhparam-2048.pem
 *.pyc
 *.sqlite3
+__pycache__
+*.sqlite3
+web/blog/migrations/
