+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+  "python.pythonPath": "C:\\Users\\Leonardo Monteiro\\AppData\\Local\\pypoetry\\Cache\\virtualenvs\\web-ttAnPv8L-py3.8\\Scripts\\python.exe",
+  "python.autoComplete.extraPaths": ["../web"],
+}
