+++ web/blog/models/BlogTag.py
@@ -1,0 +1,7 @@
+from django.db import models
+
+class BlogTag(models.Model):
+    tag = models.CharField(max_length=80)
+
+    def __str__(self):
+        return self.tag
