+++ web/blog/models/__init__.py
@@ -1,0 +1,5 @@
+from .Author import Author
+from .BlogCategory import BlogCategory
+from .BlogPost import BlogPost
+from .BlogSeries import BlogSeries
+from .BlogTag import BlogTag
