+++ web/blog/views.py
@@ -2,8 +2,8 @@ from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect
 from django.core.paginator import Paginator
 from django.utils import timezone
-from .models import BlogPost
 from .forms import BlogPostForm
+from .models.BlogPost import BlogPost
 
 
 def home_page(request):
