+++ web/blog/views/BlogCategory.py
@@ -1,0 +1,1 @@
+
