+++ web/thepacketwizards/__init__.py
@@ -1,3 +1,4 @@
 # todo: create share buttons for each blogspot
 # todo: create related posts
 # todo: create search feature
+# todo: implement Celery to post to twitter when an article is available
